Create next-on-netlify bin command
Run the `next-on-netlify` command to set up any NextJS application for 
hosting on Netlify.

It sets up a nextRouter Netlify function with SSR pages. And it copies 
HTML pages and NextJS assets to the public directory. Finally, it sets 
up the necessary Netlify _redirects to route requests to the nextRouter 
Netlify function or the HTML pages.

You must build NextJS before running `next-on-netlify`.