Add support for dynamically routed SSG pages with fallback
Copy pre-rendered, dynamically routed pages with SSG to Netlify publish
directory and copy SSG page JSON data to _next/data/ directory.

Create a Netlify Function to handle the paths not defined in the page's
getStaticPaths. When requesting a page that has not been pre-rendered,
it will be SSRed by this function. The function also returns the page
data for paths that have not been pre-defined.

See: https://github.com/FinnWoelm/next-on-netlify/issues/7