Tests: Cache NextJS builds
Cache the NextJS apps built for testing. The cache is based on the
current NextJS version and a hash of the folder contents. This
guarantees that the app will be rebuilt if the fixture files change
(e.g., modifying one of the pages or the config). Because the building
of NextJS is rather slow, the tests are several times faster when a
cache has been established.

The cache can be busted by running -npm run cleanup-.