Refactor: Split allNextJsPages.js into pages/ folder
The allNextJsPages.js file has become too long and difficult to read.
Instead, add a pages/ folder with seven subdirectories, one for each
page type of NextJS:

- api
- getInitialProps
- getServerSideProps
- getStaticProps
- getStaticPropsWithFallback
- getStaticPropsWithRevalidate
- withoutProps

Each folder contains a pages.js file that collects the relevant pages,
a setup.js file that sets up the Netlify function for the page or
copies the pre-rendered files to the correct locations, and a
redirects.js file that returns the list of redirects/rewrites to set up
for these pages.

We now have a lot more files than before, but it should be much easier
to understand than the prior allNextJsPages.js file.